backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = "Backend for a language learning application using ChatGPT."
authors = ["Tim Gushue <gushue.tim@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
Flask = "^2.0.1"
openai = "^1.3.3"
python-dotenv = "^1.0.0"
flask-cors = "^4.0.0"

[tool.poetry.dev-dependencies]
pytest = "^6.2.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
